requirements.txt
# Core Deep Learning Frameworks
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.8.0

# Medical Image Processing
monai>=1.0.0
nibabel>=3.2.0
SimpleITK>=2.1.0
pydicom>=2.3.0  # DICOM 파일 읽기

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Metrics and Evaluation
scikit-learn>=1.0.0
torchmetrics>=0.7.0

# Configuration
pyyaml>=5.4.0
omegaconf>=2.1.0

# Visualization and Logging
tensorboard>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pillow>=8.3.0

# TCIA Data Download (optional)
# tcia-utils>=0.2.0  # TCIA 데이터 프로그래밍 방식 다운로드

# Optional: For APEX mixed precision training
# nvidia-apex  # Install manually from: https://github.com/NVIDIA/apex

# Development Tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
ipython>=8.0.0
jupyter>=1.0.0
